Add files via upload
Slides have been created from README file. Hope this is along the lines you were looking for, but if not please do advise and I will alter as required.
I have pretty much copied the README as is, with some reformatting and minor corrections to the wording. 
There are some more revisions that I could make to the wording if felt necessary
Dave